pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toast-fraud-guardian"
version = "0.1.0"
description = "Real-time fraud detection and autonomous chargeback dispute system for restaurant merchants"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Redis
    "redis[hiredis]>=5.0.0",
    # Task queue
    "celery[redis]>=5.4.0",
    # ML / Inference
    "onnxruntime>=1.18.0",
    "numpy>=1.26.0",
    "torch>=2.3.0",
    # Agent workflow
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # LLM clients
    "anthropic>=0.34.0",
    "openai>=1.40.0",
    # Data validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "prometheus-client>=0.21.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=43.0.0",
    # Utilities
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "ulid-py>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
training = [
    "torch>=2.3.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "onnx>=1.16.0",
    "mlflow>=2.16.0",
]
dashboard = [
    "streamlit>=1.38.0",
    "plotly>=5.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
